[NuGet] Improve error handling when NuGet.config xml is invalid.
NuGet will use a NullSettings if it fails to load the NuGet.config
file due to invalid xml. When the user tries to save changes back
to this file the NullSettings throws an exception. Xamarin Studio
now catches this exception and shows an error message back to the
user instead of showing the full exception stack trace.

When the NuGet.config file cannot be loaded, due to invalid xml,
the user can still use the NuGet addin in Xamarin Studio but the only
package source they will be able to use is the main nuget.org feed.