[NuGet] Improve handling of invalid encryption in NuGet.config file.
Fix 19307 - Improperly configured NuGet.Config should be handled
cleanly within Xamarin Studio.

If the encrypted password in the <packageSourceCredentials> cannot
be decrypted by NuGet than an exception is thrown and not handled by
the NuGet addin.

  <packageSourceCredentials>
    <MySource>
      <add key="Username" value="MyUserName" />
      <add key="Password" value="EncryptedPasswordString" />
    </MySource>
  </packageSourceCredentials>

Invalid encryption for the password can happen if the value is edited
by hand, or if the NuGet.config file was copied from another Mac since
the encryption keys will be different.

Now if any error occurs when reading the NuGet.config, the error is
logged, and the default package source nuget.org is used instead. This
allows the user to continue to use Xamarin Studio but only with the
main NuGet package source (nuget.org).