[NuGet] Fix Web.config transforms not being applied.
Installing a NuGet package such as Nancy.Hosting.Aspnet, or
Microsoft.AspNet.WebPages would not update the web.config file in the
project with the web.config transforms included in the NuGet package.
This was because the NuGet addin was not finding the ProjectTypeGuids
and determining that the project was not a web project. Any transforms
for web.config files would then be ignored since they should only be
applied to a web project.