Fix crash when creating template with nuget pkg
The nuget package installer was relying on the current selected solution
to be set when adding the packages referenced in a template.
However, the currect solution is not always set when creating a new solution.
To fix it, a Solution argument has been added to the project template
installer extension.