[IDE] Bug fix 19422 - Unable to see Android logs according to "Filter by" option in "Android Log" window
Quick fix to handle a regression where the drop down window used by DropDownBox would not show. The drop down window being a TopLevel (which was done to support the path bread crumb) causes a focus issue with makes the DropDownBox close the window. There is a focus handler in DropDownBox that destroys the drop down window, and removing that is also an option, but using a Popup works better. A better approach needs to be investigated, similar to combo box with a gtk grab.