[NuGet] Add unit tests for XDT fixes.
Add tests to check the changes made to the Xml Document Transformation
library and verify it works on Mono when running Remove or
SetAttributes transforms.

This should catch if the wrong Microsoft.Web.XmlTransform.dll is used.

These tests should be removed once the XDT library is being built from
source code along with NuGet. The XDT library contains its own set of
unit tests that can be run on Mono which check these transforms are
working.