[NuGet] Support installing specific package version.
The Add Packages dialog can now search for all versions of a package.

Search syntax supported:

    PackageId -version
    PackageId -v
    PackageId -ver

This returns all versions of the package ordered with the latest
version first in the list.

The package id must exactly match the package id otherwise no results
will be returned.

In the list of packages the package version is displayed instead of
the download count when a search is run for multiple package versions.
This makes it easier to see the versions without having to select each
package in the list. The download count is displayed for all other
searches.