[Ide] Ensure that the editor is initialised when changing views in the workspace window.
If a view has an additional view that implements IExtensibleTextEditor, ensure that it is initialised. Some designer views are the primary view and the editor does
not get a chance to be initialised when the document is attached and can only be initialised when the "source" editor becomes the active view.