[NuGet] Reduce packages.config file reads.
When the target framework is changed in project options the
ProjectModified event is raised twice. Once for the project's
BaseDirectory and once for the projects' TargetFramework. The
ProjectModified event then causes the Packages node to be reloaded.
Each Packages node refresh would cause two separate reads of the
packages.config file. The results of the first read are now cached
until the Packages node is refreshed.

This logic can be improved further to prevent multiple reads of
the packages.config file when nothing has changed. This would require
the current state of a project's package references to be cached
instead of relying on the packages.config file. Or a way to determine
that the packages.config file has been updated and should be re-read.