[NuGet] Fix access denied error when target framework changed.
Fix UnauthorizedAccessException: Access to the path
'd:\Test\packages.config' is denied. This would occur when the target
framework of the project is changed to one that requires the
project's NuGet packages to be reinstalled.

The Solution pad is updating the Packages folder by reading the
packages.config file at the same time that the compatibility checker
is checking the NuGet packages are compatible in the background. When
the packages.config file is updated with requireReinstallation="True"
this can occasionally fail due to the file being read at the same
time. Short term fix is to update the packages.config on the UI
thread. Longer term fix is to use better logic when the Solution pad
needs to update the Packages node.