[Ide] Allow search data source to access the original search text.
Allow the original text used in the unified search text box to be
accessed by those search categories that may require it. If the
search uses ':' in the text then it is split up into a tag and a
pattern. With the text separated it can be difficult and sometimes
impossible to recreate the original search text. This is a problem for
those search categories that allow the use of ':' as part of the
search.