[NuGet] Fix unified search when NuGet search tags are used.
Previously package searches that used a ':' character would search for
the wrong thing when launched from the unified search. This is because
the unified search text is parsed and split up into a tag and a search
pattern based on the ':' character. NuGet package searches allow the
':' character but the search that was being used contained only part
of the original search text so the search would be incorrect.

Now the full original search text is used when Search Packages is
selected.

So now the following NuGet package searches will now work when
executed from the unified search:

    id:xamarin.forms
    packageid:xamarin.forms
    tags:typescript