[Ide] Add support for cross-toolkit images
Xwt.Drawing.Image objects can now reference native images from different
toolkits. For example, a single Image object can contain a reference
to a Gdk.Pixbuf and an NSImage. The goal is to be able to use a single
object instance to represent an image that can be rendered in any
supported toolkit.

ImageService now loads the icons using a special api that allows
specifying the image source. This source will be used to load
platform-specific images on demand.