[NuGet] Fix second Package Console window being created.
On opening a solution a second Package Console window would be created.

The check for package updates would start whilst the
"NuGet.exe restore" was still running resulting in two Package Console
windows. Now the package restorer waits until "NuGet.exe restore" is
completed to ensure the check for package updates happens afterwards.

It is planned to stop using "NuGet.exe restore" and move the logic
into the NuGet addin so the restore can prompt for credentials if they
are missing instead of failing to restore.