[NuGet] Fixed Bug 18951 - Incorrect path separator used for Import.
https://bugzilla.xamarin.com/show_bug.cgi?id=18951

Incorrect path separator used when adding MSBuild .targets/.props
file to a project when installing a NuGet package. On the Mac the
forward slash character was being used for the path to the .targets
file. All paths in MSBuild project files should use backslash on
all platforms. If forward clashes are used then updating a NuGet
package on Windows will fail to remove the old import from the
project. It will also cause the build to fail since the path is
resolved to a directory inside Program Files.