[Core] Allow a Condition to be set for an MSBuild Import.
A condition specified in the call DotNetProject.AddImportIfMissing(..)
will now be saved in the MSBuild project file. This allows imports to
be conditional.

One use is for NuGet packages that define a custom MSBuild targets
file and allows the targets file to be conditionally imported if the
file exists.

<Import
  Project="packages\Xamarin.Forms\Xamarin.Forms.targets"
  Condition="Exists('packages\Xamarin.Forms\Xamarin.Forms.targets')"/>

Using the Exists check allows Visual Studio to open the project when
the package is missing and has not yet been restored.