[NuGet] Fix hang when restoring packages from menu.
To allow the package update check to start after the package restore
had finished the logic was changed so it waited until NuGet.exe
restore had completed. When running a restore from the menu the wait
occurred on the UI thread causing a hang. The restore is now run on a
background thread in the same way the restore occurs on opening a
solution.