[NuGet] Move cursor to end when search started from unified search.
Previously the cursor would be at the start of the search text in the
Add Packages dialog when a uGet package search was started from the
unified search. Now the cursor is at the end of the search text at the
top right of the dialog as though the text was typed directly into the
Add Packages dialog.