[NuGet] Disable Add Package button whilst typing in search text.
When a package was checked the Add Package button was enabled whilst
the search text was being typed in. With the new logic that allows a
package to be checked but not visible on screen the Add Package button
was incorrectly enabled whilst the search text was being typed in. Now
if the loading message is being displayed the Add Package button is
disabled.