[NuGet] Enter key in Add Packages should install checked packages.
When the search text box has the focus with one or more packages were
checked and the enter key was pressed instead of the checked packages
being installed the currently selected package was installed. This was
inconsistent with what happens when the enter key is pressed whilst
the list view has the focus.

Pressing enter whilst the search text box has focus is supposed to be
equivalent to pressing the Add Package button.

Now when the search text box has focus and there are one more checked
packages then pressing enter will result in the checked packages being
installed. Pressing the enter key during a search will not install any
packages whilst the loading message is displayed.