[NuGet] Restore packages using -NonInteractive command line option.
Without this command line option using NuGet.exe restore with a
package source that requires credentials that are not saved in the
NuGet.config file will result in the Package Console trying to prompt
for credentials. On the Mac you will see the following in the Package
Console output:

  UserName: Password: ***************************WARNING: > 65536
  Parameter name: length
  WARNING: > 65536

NuGet.exe does not stop but prints out the above output and then
fails. Using the -NonInteractive option a warning message will be
displayed and then an error message will be displayed about being
unable to find the NuGet package:

  Please provide credentials for: http://mynugetfeed.com/
  UserName: WARNING: Cannot prompt for input in non-interactive mode.
  Unable to find version '5.0.7' of package 'Newtonsoft.Json'.

This is a short term fix. Long term fix is to move the NuGet restore
logic into the NuGet addin so the user is prompted to enter their
username and password into the credentials dialog.