[NuGet] Handle unrestored packages when looking for package updates.
Previously when checking for updates the NuGet addin would only look
for package updates for restored packages. So if the Automatic package
restore option was disabled you could see a status bar message
indicating that there were no package updates when actually there were.
This is because the NuGet addin would find no packages for the project
since they were unrestored and not look for any updates.

Now the NuGet addin uses the package references defined in the
packages.config file. This means that even if the packages are not
restored then package updates can be found.

There is still the problem that you cannot update a package if it is
restored. The NuGet addin should check if the package is missing
when an update is attempted and then restore the packages first.