[NuGet] Fix incorrect paths added to repositories.config on Mac.
With a custom repositoryPath defined in the solution level
NuGet.config file the paths written to the repositories.config
had an extra ../ at the start. These paths being wrong meant that
a NuGet package installed from one project in a solution would
result in it being removed from the packages directory even if
another project was referencing it.