[NuGet] Add Packages dialog -  Package sources could not be reached.
When All Sources is selected and if any one of the package sources
threw an exception then no packages were returned. Now as long as one
package source returns packages then the dialog will show the packages
and a warning indicating that some package sources could not be
reached is displayed as well as the errors for the failing package
sources. This warning text is at the top of the package list with an
orange background.

If all package sources throw an exception then the warning message
indicates that no package source could be reached along with the error
messages for each package source.

In order to support this feature two small modifications were made to
NuGet's AggregateRepository. One modification to support aggregate
package sources still returning packages even if one package source
failed. The second change was to to allow access to the failing
repositories outside of the AggregateRepository class.