[NuGet] Add Packages dialog - Clear error when search text entered.
When search text is typed into the text box any error or warning
message is now cleared. Previously just the packages were removed and
the error message was still displayed, then when the new packages were
returned the error message would be removed and a new error message
would be displayed again. This would cause the UI in the package list
to flicker.

Clearing the error message should indicate that the new search has the
an error and the error message is not related to the previously
displayed error.