[NuGet] Fix empty source selected in Add Packages dialog.
An empty package source selected in the Add Packages dialog could
occur when two enabled package sources and All Sources selected in the
Add Packages dialog. Then Configure Sources is selected and one of the
enabled package sources is unchecked in Preferences. Then on
opening the Add Packages dialog again an empty package source is
displayed as the selected package source.