[NuGet] Add Packages dialog - Fix warning displayed for valid sources.
The warning about unreachable package sources was being displayed in
the Add Packages dialog even though the invalid package source was
disabled. This could happen if the Add Packages was opened once with
All Sources selected and one failing package source, then Configure
Sources is selected, in Preferences the failing package source is
disabled and another package source which is valid is enabled. Then
on switching back to the Add Packages dialog once the packages are
returned from the valid package sources the original warning was
displayed since it was being cached in the aggregate repository. Now
when the package sources are changed in Preferences the aggregate
repository is recreated which gets rid of any cached failing
repositories and warnings.