[Ide] Drag notebook fixes
Apply a tolerance marging when dragging a tab.
This fixes bug #20991 - On dragging and switching tab positions, smaller 
version of the document/view keeps appearing with overlay

Cancel the drag&drop operation if the grab on the notebook is broken.
This fixes bug #20989 - Smaller version of the document/view being dragged
sticks to mouse pointer, if user switches to some other window