[NuGet] Fix updated package version missing for unrestored packages.
Package updates are checked even if automatic package restore is
disabled. Previously if the packages were not restored, and there are
updated packages for a project when it was opened, the number of
updated packages was shown on the Packages folder in the Solution
window but the updated package version was not displayed.

Now unrestored packages and packages requiring retargeting will be
displayed in the Solution window with orange text and a warning
icon but will also have the updated package version number displayed.