[NuGet] Show packages added to solution in Add Packages dialog.
Opening the Add Packages dialog will show the packages added to all
projects in the current solution.

When a package is installed this will appear at the top of the list
of packages in the Add Packages dialog. List order is recent packages,
solution packages, packages from active package source.