[NuGet] Restore packages for single project when updating that project.
Fixed bug 21125 - On updating any single project packages/Package,
all packages of that solution get restored.
https://bugzilla.xamarin.com/show_bug.cgi?id=21125

The NuGet addin now restores packages for the selected project instead
of for the entire solution. This affects the following actions:

1) Right click a package reference in the Packages folder and select
Update. This will now restore packages for the project that
has the package reference if that project has any missing packages.
Note that here the NuGet addin will not just restore the single
package that is selected. The reason for this is that this package may
depend on other packages which may need to be updated at the same
time. In theory it would be possible to just restore the minimal set
of packages by restoring just one, then checking its dependencies,
then restoring another, but it seems simpler and less error prone to
just to restore all the packages for the project in this case.

2) Right click the Packages folder and select Update. This will now
restore only those packages for the project if they are missing before
updating the packages.