[NuGet] Do not check for package updates if project has no packages.
Fixed bug 21194 - Message "Packages are up to date." is shown in
Status bar and Package Console window while project does not have any
package.
https://bugzilla.xamarin.com/show_bug.cgi?id=21194

If automatic package restore was disabled, and check for updates was
enabled, the NuGet addin would check for package updates when a
solution was opened. This would result in the
"Packages are up to date" message being displayed in the status bar.

Now if the project has no packages.config file then the NuGet addin
will not check for updates so the status bar message will not appear.