[NuGet] Remove duplicate warnings reported at end of Package Console.
Fixed bug 21592 - Message 'No Update found but warnings were reported'
is displaying multiple times in Package Console Output window.
https://bugzilla.xamarin.com/show_bug.cgi?id=21592

When a package source was unavailable and the user tried to update
multiple packages in the project then for each update there would be
a warning displayed at the end of the Package Console output for each
failure to reach the package source.

Warnings are displayed at the end of the Package Console so you do not
have to scroll through the console to find them.

Now the NuGet addin will not show these duplicate warnings at the end
of the Package Console output.