[NuGet] Show warning for invalid source after checking for updates.
Fixed bug 21629 - With invalid package source, when open any
project containing no updated packages, it does not display a warning.
https://bugzilla.xamarin.com/show_bug.cgi?id=21629

Previously when one or more package source was unreachable or invalid
the NuGet addin would show a "Packages are up to date." message in the
status bar after checking for package updates even if this may not be
true since not all package sources could be reached.

Now if a warning is logged when checking for package updates the
status bar now shows the message "No updates found but warnings were
reported. Please see Package Console for details."

Warnings from invalid package sources are now logged in the Package
Console when checking for updated packages.