[NuGet] Provide package add/remove events for components integration.
Extend the IPackageManagementProjectOperations interface so it has
events for a package being added or removed. Note that these events
are only fired if a package is selected in the Add Packages dialog
and added, or if the package is selected in the Solutions window and
removed. Events will not fire for any package dependencies that are
added or removed. This should be OK for components since the events
are being used as a way to detect if the NuGet packages in a project
have changed and not if a particular NuGet package was added or
removed.