[NuGet] Fix Packages folder not showing items during Component install
When a Xamarin Component that has NuGet package dependencies was being
installed the Packages folder would not show any packages until the
installation finished. This was caused by a null reference exception
in the Packages node builder since it was assuming that a full
IPackage was available in the pending package install actions. However
for a Xamarin Component installation only the package id and package
version are available. The install package action allows either the
IPackage to be specified or the package id and version. To fix this
the Packages folder node builder asks the install package action for
the package id and version, which is taken from either the IPackage
or the properties on the action itself.