[NuGet] Use version constraints when checking for package updates.
A version constraint defined in the packages.config file, as shown
below, is now used when checking for package updates from a package
source. Previously always the latest version of the NuGet package
would be displayed as an update in the Solution window.

  <package
    id="jQuery"
    version="1.4.1"
    targetFramework="net40"
    allowedVersions="[1.4.1,1.8)" />

With the above constraint when the 'Check for package updates on
opening a solution' option is enabled in preferences then only jQuery
1.7.2 is shown as available update in the Solutin window.

Previously the latest version of jQuery (currently 2.1.1) would be
shown as an available update. After updating to jQuery 1.7.2 then no
further updates would be shown as available in the Solution window.