[build] Change the location of the 'lib' directory
This contains non-executable files only (according to apples
logic). As such we should not put it in the 'MacOS' directory
in the app bundle as new signing rules require every file in
there to be signed and apple also explicitly state this directory
should only contain signed code.

We also can't sign the contents of the 'lib' directory because
apples tools get confused by the dots in the directory names
and assume that they are frameworks instead of regular directories.
This causes signing to fail because the contents of our directories
do not contain the expected 'MacOS' and 'Resources' directories.