[Core] Fix shared assets project loading issue
In some cases the name of the .projitems files may not match the name
of the .shaproj file. This fix gets the name of the .projitems from the
content of the .shproj, instead of just changing the extension.

Fixes bug #20571 - References to Shared Asset Projects are forgotten in
between xam studio restarts.