[Ide] Fix focus issues when showing pads as NSWindow children.
Work around the focus issues by switching to WindowType.Popup and showing the window
*after* we add child. However, realizing the window before it's shown causes it to
appear at the display origin and then get moved into the desired location when we
show it. This causes a highly visible flickering issue. There isn't any apparent want
to prevent this, so we're masking the flickering using an opacity trick.